Reorganize component files into folders
