+++ src/pages/education.js
@@ -1,8 +1,8 @@
 import React from 'react'
-import MainLayout from '../components/mainlayout'
+import MainLayout from '../components/layouts/mainlayout'
 
 export default function Education() {
     return (
-        <MainLayout title='Education' selectedTab='education' />
+        <MainLayout title='Education' selectedTab='education'/>
     )
 }
